refactor: 전체적인 디자인 수정
