+++ src/components/InformModal/InformModal.module.css
@@ -21,7 +21,8 @@
   border-radius: 17.902px;
   background: #fff;
 }
-span {
+.InformText,
+.Graytext {
   color: #000;
   text-align: center;
   font-family: Pretendard;
