--- src/pages/Signup.tsx
@@ -1,10 +1,0 @@
-// 회원가입 페이지
-import React from "react";
-
-type Props = {};
-
-const Signup = (props: Props) => {
-  return <div>Signup</div>;
-};
-
-export default Signup;
